+++ src/main/java/com/knighten/ai/search/AbstractAStarNode.java
@@ -13,17 +13,17 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
     /**
      * Node's h() heuristic value.
      */
-    private int hScore = Integer.MIN_VALUE;
+    private double hScore = Integer.MIN_VALUE;
 
     /**
      * Node's g() graph cost value.
      */
-    private int gScore = 0;
+    private double gScore = 0;
 
     /**
      * Node's f() total score value. This is h() + g().
      */
-    private int fScore = 0;
+    private double fScore = 0;
 
     /**
      * Node's parent.
@@ -35,14 +35,13 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
      */
     private T state;
 
-
     /**
      *  Returns the heuristic score h() for the node. You MUST call calcH() before attempting to retrieve the heuristic
      *  score, otherwise h() = Integer.MIN_VALUE.
      *
      * @return the heuristic score h() for the node
      */
-    public int getH() {return this.hScore;}
+    public double getH() {return this.hScore;}
 
     /**
      * Sets the node's heuristic score h(). Get h() from A IHeuristicFunction object then store the value using this
@@ -50,21 +49,21 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
      *
      * @param hScore the heuristic score h() of the node
      */
-    public void setH(int hScore) {this.hScore = hScore;}
+    public void setH(double hScore) {this.hScore = hScore;}
 
     /**
      * Returns the graph link score g(). The total path cost accumulated so far to reach current node.
      *
      * @return  returns graph link score g()
      */
-    public int getG() {return this.gScore;}
+    public double getG() {return this.gScore;}
 
     /**
      * Set the node's path cost g(). The total path cost accumulated so far to reach current node.
      *
      * @param  gScore the node's path cost g()
      */
-    public void setG(int gScore) {this.gScore = gScore;}
+    public void setG(double gScore) {this.gScore = gScore;}
 
     /**
      * Gets the node's parents. Initial nodes should have parent equal to NULL.
@@ -86,14 +85,14 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
      *
      * @return the f() value of the node
      */
-    public int getF() {return this.fScore;}
+    public double getF() {return this.fScore;}
 
     /**
      * Sets the node's f() value.
      *
      * @param fScore the f() value of the node
      */
-    public void setF(int fScore) {this.fScore = fScore;}
+    public void setF(double fScore) {this.fScore = fScore;}
 
     /**
      * Returns the node's state
@@ -117,7 +116,9 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
      * @return  negative if node1 LT node2; zero if node1==node2; positive if node1 GT node2;
      */
     @Override
-    public int compare(AbstractAStarNode<T> node1, AbstractAStarNode<T> node2) {return node1.getF() - node2.getF();}
+    public int compare(AbstractAStarNode<T> node1, AbstractAStarNode<T> node2) {
+        return (int) (node1.getF() - node2.getF());
+    }
 
     //////////////////////
     // Abstract Methods //
@@ -135,7 +136,7 @@ public abstract class AbstractAStarNode<T> implements Comparator<AbstractAStarNo
      *
      * @return returns distance to node's parent
      */
-    public abstract int distFromParent();
+    public abstract double distFromParent();
 
     /**
      * Generates the hash code for the node. The hash should be based on node state. We want nodes with the same
